requirements.txt
Flask==2.3.4
pycryptodome==3.19.0
opencv-python==4.8.1.78
numpy==1.26.0
pyserial==3.7
face_recognition==1.3.0
mysql-connector-python==8.1.1
